+++ dwds/test/chrome_proxy_service_test.dart
@@ -578,9 +578,104 @@ void main() {
       expect(version.major, greaterThan(0));
     });
 
-    test('invoke', () {
-      expect(() => service.invoke(null, null, null, null),
-          throwsUnimplementedError);
+    group('invoke', () {
+      VM vm;
+      Isolate isolate;
+      InstanceRef testInstance;
+
+      setUp(() async {
+        vm = await service.getVM();
+        isolate = await service.getIsolate(vm.isolates.first.id);
+        testInstance = await service.evaluate(
+            isolate.id, isolate.rootLib.id, 'myInstance') as InstanceRef;
+      });
+
+      test('toString()', () async {
+        var remote =
+            await service.invoke(isolate.id, testInstance.id, 'toString', []);
+        expect(
+            remote,
+            const TypeMatcher<InstanceRef>().having(
+                (instance) => instance.valueAsString,
+                'toString()',
+                "Instance of 'MyTestClass'"));
+      });
+
+      test('hello()', () async {
+        var remote =
+            await service.invoke(isolate.id, testInstance.id, 'hello', []);
+        expect(
+            remote,
+            const TypeMatcher<InstanceRef>().having(
+                (instance) => instance.valueAsString, 'hello()', 'world'));
+      });
+
+      test('helloString', () async {
+        var remote = await service.invoke(isolate.id, isolate.rootLib.id,
+            'helloString', ['#StringInstanceRef#abc']);
+        expect(
+            remote,
+            const TypeMatcher<InstanceRef>().having(
+                (instance) => instance.valueAsString, 'helloString', 'abc'));
+        expect(
+            remote,
+            const TypeMatcher<InstanceRef>()
+                .having((instance) => instance.kind, 'kind', 'String'));
+      });
+
+      test('null argument', () async {
+        var remote = await service.invoke(
+            isolate.id, isolate.rootLib.id, 'helloString', ['objects/null']);
+        expect(
+            remote,
+            const TypeMatcher<InstanceRef>().having(
+                (instance) => instance.valueAsString, 'helloString', 'null'));
+        expect(
+            remote,
+            const TypeMatcher<InstanceRef>()
+                .having((instance) => instance.kind, 'kind', 'Null'));
+      });
+
+      test('helloBool', () async {
+        var remote = await service.invoke(
+            isolate.id, isolate.rootLib.id, 'helloBool', ['objects/bool-true']);
+        expect(
+            remote,
+            const TypeMatcher<InstanceRef>().having(
+                (instance) => instance.valueAsString, 'helloBool', 'true'));
+        expect(
+            remote,
+            const TypeMatcher<InstanceRef>()
+                .having((instance) => instance.kind, 'kind', 'Bool'));
+      });
+
+      test('helloNum', () async {
+        var remote = await service.invoke(
+            isolate.id, isolate.rootLib.id, 'helloNum', ['objects/int-123']);
+        expect(
+            remote,
+            const TypeMatcher<InstanceRef>().having(
+                (instance) => instance.valueAsString, 'helloNum', '123'));
+        expect(
+            remote,
+            const TypeMatcher<InstanceRef>()
+                .having((instance) => instance.kind, 'kind', 'Double'));
+      });
+
+      test('two object arguments', () async {
+        var remote = await service.invoke(isolate.id, isolate.rootLib.id,
+            'messagesCombined', [testInstance.id, testInstance.id]);
+        expect(
+            remote,
+            const TypeMatcher<InstanceRef>().having(
+                (instance) => instance.valueAsString,
+                'messagesCombined',
+                'worldworld'));
+        expect(
+            remote,
+            const TypeMatcher<InstanceRef>()
+                .having((instance) => instance.kind, 'kind', 'String'));
+      });
     });
 
     test('kill', () {