+++ dwds/test/inspector_test.dart
@@ -4,6 +4,7 @@
 
 @TestOn('vm')
 import 'package:dwds/src/connections/debug_connection.dart';
+import 'package:dwds/src/utilities/conversions.dart';
 import 'package:dwds/src/debugging/inspector.dart';
 import 'package:dwds/src/utilities/shared.dart';
 import 'package:test/test.dart';
@@ -36,12 +37,12 @@ void main() {
   String libraryVariableExpression(String variable) =>
       '$loadModule("dart_sdk").dart.getModuleLibraries("web/scopes_main")["$url"]["$variable"];';
 
-  Future<RemoteObject> libraryPublicFinal() => inspector
-      .evaluateJsExpression(libraryVariableExpression('libraryPublicFinal'));
+  Future<RemoteObject> libraryPublicFinal() =>
+      inspector.jsEvaluate(libraryVariableExpression('libraryPublicFinal'));
 
   test('send toString', () async {
     var remoteObject = await libraryPublicFinal();
-    var toString = await inspector.sendMessage(remoteObject, 'toString', []);
+    var toString = await inspector.invokeMethod(remoteObject, 'toString', []);
     expect(toString.value, 'A test class with message world');
   });
 
@@ -71,9 +72,82 @@ void main() {
       'count',
       'message',
       'myselfField',
-      'notFinal'
+      'notFinal',
+      'tornOff',
     ];
     names.sort();
     expect(names, expected);
+  });
+
+  group('invoke', () {
+    // We test these here because the test fixture has more complicated members to
+    // exercise.
+
+    String isolateId;
+    String libraryId;
+    RemoteObject instance;
+
+    setUp(() async {
+      isolateId = inspector.isolate.id;
+      libraryId = inspector.isolate.rootLib.id;
+      instance =
+          await inspector.evaluate(isolateId, libraryId, 'libraryPublicFinal');
+    });
+
+    test('invoke top-level private', () async {
+      var remote = await inspector.invoke(isolateId, libraryId,
+          '_libraryPrivateFunction', [dartIdFor(2), dartIdFor(3)]);
+      expect(
+          remote,
+          const TypeMatcher<RemoteObject>()
+              .having((instance) => instance.value, 'result', 5));
+    });
+
+    test('invoke instance private', () async {
+      var remote = await inspector.invoke(isolateId, instance.objectId,
+          'privateMethod', [dartIdFor('some string')]);
+      expect(
+          remote,
+          const TypeMatcher<RemoteObject>().having((instance) => instance.value,
+              'result', 'some string : a private field'));
+    });
+
+    test('invoke instance method with object parameter', () async {
+      var remote = await inspector
+          .invoke(isolateId, instance.objectId, 'equals', [instance.objectId]);
+      expect(
+          remote,
+          const TypeMatcher<RemoteObject>()
+              .having((instance) => instance.value, 'result', true));
+    });
+    test('invoke instance method with object parameter 2', () async {
+      var libraryPrivateList =
+          await inspector.evaluate(isolateId, libraryId, '_libraryPrivate');
+      var remote = await inspector.invoke(isolateId, instance.objectId,
+          'equals', [libraryPrivateList.objectId]);
+      expect(
+          remote,
+          const TypeMatcher<RemoteObject>()
+              .having((instance) => instance.value, 'result', false));
+    });
+
+    test('invoke closure stored in an instance field', () async {
+      var remote =
+          await inspector.invoke(isolateId, instance.objectId, 'closure', []);
+      expect(
+          remote,
+          const TypeMatcher<RemoteObject>()
+              .having((instance) => instance.value, 'result', null));
+    });
+
+    test('invoke a torn-off method', () async {
+      var toString = await inspector.loadField(instance, 'tornOff');
+      var result =
+          await inspector.invoke(isolateId, toString.objectId, 'call', []);
+      expect(
+          result,
+          const TypeMatcher<RemoteObject>().having((instance) => instance.value,
+              'toString', 'A test class with message world'));
+    });
   });
 }